Coords

Description

Documentation

type RatCoord = (Rational, Rational)

Type alias for rational coordinates

type IntCoord = (Int, Int)

Type alias for integer coordinates

addCoord

Arguments

:: Num a

=> (a, a)
First point

-> (a, a)
Second point

-> (a, a)
Result of the translation

Translates the first point using the second one as a movement vector.

toIntCoord

Arguments

:: RatCoord
Rational coordinates to round

-> IntCoord
Result of the rounding

Rounds the point coordinates to the nearest integers to give a IntCoord