Game

Description

Documentation

type KeyMapping = Key -> Maybe ActionType

Type alias representing the key mapping used by the game

data ActionType

Either an in-game move action, or a control action (play, pause, reset...)

Constructors

Move MoveType

Control ControlType

data MoveType

All possible moves in the game

Constructors

RotateLeft
Counter-clockwise rotation

RotateRight
Clockwise rotation

MoveDown
Move down (1 step)

MoveLeft
Move left (1 step)

MoveRight
Move right (1 step)

FullDrop
Move down (as much as possible)

data ControlType

All control actions (not related to in-game movement)

Constructors

PauseResume
Pause or resume the current game

Reset
Reset the game

ticksPerSecond :: Int

Controls the update rate of the game screen and data

initialSpeed :: Float

Controls the speed at which pieces fall at the beginning of the game

dropTickInterval

Arguments

:: Float
Current speed factor

-> Int
Resulting drop tick interval

Computes the interval between two drop ticks using ticksPerSecond and the current speed

playableRowNb :: Int

Number of rows in the grid that the player can actually use

totalRowNb :: Int

Number of rows in the grid including the two where pieces spawn

colNb :: Int

Number of columns in the grid

isInPlayableGrid :: IntCoord -> Bool

Tells whether a 2D-point is in the playable part of the grid or not

isInTotalGrid :: IntCoord -> Bool

Tells whether a 2D-point is in the grid, including the spawn zone, or not.

isInGridOrAbove :: IntCoord -> Bool

Tells whether a 2D-point is in the grid and its vertical extension, or not.

defaultKeyMapping :: KeyMapping

Default key mapping, based on FPS controls