Grid

Description

Documentation

type Grid a = Vector (Vector a)

Type alias for the row-based grid

(!!!) infixl 9

Arguments

:: Grid a
Grid where the value will be fetched

-> IntCoord
2D index (integer coordinates) indicating where to fetch the value in the grid

-> a
The resulting value

Implements 2D indexing for grid. Fails if the 2D index is out of bounds.

fullRows

Arguments

:: Grid (Maybe a)
Grid to work on

-> [Int]
indices (0-based) of rows which only contains Just values

When the grid is composed of Maybe values, returns the indices of rows which only contain Just values.

deleteRows

Arguments

:: Grid a
Grid to work on

-> [Int]
Indices of the rows to delete

-> a
Default value which will be replicated to create new rows

-> Grid a
Resulting grid

Removes the rows whose indices are given, push the other rows closer to the bottom of the grid, and complete the grid with new rows at the top, created using the specified default value.

update

Arguments

:: Grid a
Grid to work on

-> [(IntCoord, a)]
Pairs of (2D-index, value) describing the updates to make

-> Grid a
Resulting grid

Updates the given grid using the (2D-index, value) pairs.

mapGrid

Arguments

:: (a -> b)
Function to apply on each cell of the grid

-> Grid a
Grid to work on

-> Grid b
Resulting grid

Applies the specified function to each cell of the grid. (I could not implement fmap for 'Grid a' because 'Grid a' is only a type alias)

lMapGrid

Arguments

:: (IntCoord -> a -> b)
Function to apply on each (2D-index, value) pair

-> Grid a
Grid to work on

-> [b]
Resulting list

Applies the specified function to each pair (2D-index, value) of the grid, and collects all the results in a list.