Pieces

Description

Documentation

data Orientation

Represents the four orientations a piece can take in Tetris

Constructors

S0
Initial state, horizontal (for most pieces)

S1
After 1 clockwise rotation action

S2
After 2 rotation actions

S3
After 3 clockwise rotation actions

data PieceType

Tetrominos types

Constructors

I

J

L

O

S

T

Z

data MovingPiece

The moving piece (i.e. one which is still falling) is defined by its type, orientation and center of rotation

Constructors

MovingPiece

Fields

pieceType :: PieceType

Type of tetromino

orientation :: Orientation

Current orientation of the piece

center :: RatCoord

Center of rotation of this piece. Can be a rational point (we consider that the center of each grid cell has integer coordinates, so the borders/corners of cells have rational coordinates)

type PieceGrid = Grid (Maybe PieceType)

Type alias for a grid in which each cell can contain a colored square or be empty.

generatePieceBag :: IO [PieceType]

Generates a list of piece types where each type appears only one (but in a random order).

spawnPiece

Arguments

:: PieceType
Type of tetromino to spawn

-> MovingPiece
The resulting moving piece

Creates a new moving piece using the specified type.

applyMove

Arguments

:: PieceGrid
The grid to work on. Used to check which moves are valid or not

-> MovingPiece
The current state of the falling piece

-> MoveType
Move requested by the user

-> Maybe MovingPiece
Updated moving piece if the move can be executed

Given the grid, the current moving piece and the move type requested by the user, returns a moving piece with its orientation and position updated, if the move is possible.

solidifyPiece

Arguments

:: PieceGrid
The grid to update

-> MovingPiece
The moving piece to solidify

-> PieceGrid
The updated grid

When a moving piece has finished falling, this function transforms it into colored square cells in the grid.

selfRepr

Arguments

:: MovingPiece
The current moving piece

-> [IntCoord]
List of grid coordinates on which to show the piece

Returns the grid cells on which the moving piece must be shown.

ghostRepr

Arguments

:: PieceGrid
The grid

-> MovingPiece
The current moving piece

-> [IntCoord]
List of grid coordinates on which to show the piece "ghost"

Returns the grid cells on which the "ghost" (i.e. the ground projection) of the moving piece must be shown.

isValid

Arguments

:: PieceGrid
The grid to work on

-> MovingPiece
The piece position to check

-> Bool
True iif the moving piece position is valid

Tells whether a moving piece position is valid or not: + all its squares must be in the grid or its vertical extension + the squares must not overlap an already solidified colored square in the grid