World

Description

Documentation

data World

Represents the game state

Constructors

World

Fields

keyMapping :: KeyMapping

Mapping from keyboard keys to game actions

grid :: PieceGrid

Grid of solidified pieces (i.e. pieces which are no longer falling)

movingPiece :: MovingPiece

The tetromino which is currently falling

pieceBag :: [PieceType]

Next piece types to spawn

rowsToDelete :: [Int]

Completed rows, which will be deleted next tick

currentSpeed :: Float

Current speed of the game. Determines the tick interval between two "gravity" moves

tickNo :: Integer

Tick count

tickNextDrop :: Integer

Tick number where the next "gravity" move will happen

paused :: Bool

Is the game paused?

finished :: Bool

Is the game over?

rowsCleared :: Int

Number of complete rows cleared by the player

makeInitialWorld :: KeyMapping -> IO World

Creates the initial game state (using the specified key mapping).

updateWorldWithEvent

Arguments

:: Event
The received event

-> World
Current game state

-> IO World
Next game state

Updates the game state when an event is issued.

updateWorldWithTick

Arguments

:: Float
Number of seconds elapsed since the last call to this function

-> World
Current game state

-> IO World
Next game state

Updates the game state when a tick is elapsed